# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plannify Driver openapi",
    "description": "API documentation for the Plannify Driver API",
    "contact": {
      "name": "contact@plannify.be"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api-dev.plannify.be/driver/v1"
    },
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/me": {
      "get": {
        "tags": [
          "driver"
        ],
        "description": "Get driver personal information.",
        "operationId": "get_driver_info",
        "responses": {
          "200": {
            "description": "Driver information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDriverResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "driver"
        ],
        "description": "Update driver personal information. Only provided fields are updated.",
        "operationId": "update_driver_info",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Driver information updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDriverResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Driver not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Email already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/me/deactivate": {
      "post": {
        "tags": [
          "driver"
        ],
        "description": "Deactivate the authenticated driver's account.",
        "operationId": "deactivate_driver",
        "responses": {
          "200": {
            "description": "Account deactivated successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Account is already deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/me/reactivate": {
      "post": {
        "tags": [
          "driver"
        ],
        "description": "Reactivate the authenticated driver's account.",
        "operationId": "reactivate_driver",
        "responses": {
          "200": {
            "description": "Account reactivated successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Account is not deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/rest-periods": {
      "get": {
        "tags": [
          "driver/rest-periods"
        ],
        "description": "Retrieve driver rest periods",
        "operationId": "get_all_rest_periods",
        "responses": {
          "200": {
            "description": "Driver rest periods retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverRestPeriod"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "driver/rest-periods"
        ],
        "description": "Set driver rest periods",
        "operationId": "set_rest_periods",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverRestPeriodsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Driver rest periods set successfully"
          },
          "400": {
            "description": "Invalid rest period",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "driver/rest-periods"
        ],
        "description": "Delete driver rest periods",
        "operationId": "delete_rest_periods",
        "responses": {
          "200": {
            "description": "Driver rest periods deleted successfully"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDriverResponse": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          }
        }
      },
      "CreateDriverRestPeriodRequest": {
        "type": "object",
        "required": [
          "start",
          "end",
          "rest"
        ],
        "properties": {
          "end": {
            "type": "string"
          },
          "rest": {
            "type": "string"
          },
          "start": {
            "type": "string"
          }
        }
      },
      "CreateDriverRestPeriodsRequest": {
        "type": "object",
        "required": [
          "rest_periods"
        ],
        "properties": {
          "rest_periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateDriverRestPeriodRequest"
            }
          }
        }
      },
...
      "DriverLanguage": {
        "type": "string",
        "enum": [
          "fr",
          "en"
        ]
      },
      "DriverRestPeriod": {
        "type": "object",
        "required": [
          "start",
          "end",
          "rest"
        ],
        "properties": {
          "end": {
            "type": "string"
          },
          "rest": {
            "type": "string"
          },
          "start": {
            "type": "string"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "message",
          "error_code",
          "status"
        ],
        "properties": {
          "content": {},
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "GetDriverResponse": {
        "type": "object",
        "required": [
          "pk_driver_id",
          "firstname",
          "lastname",
          "email",
          "is_searchable",
          "allow_request_professional_agreement",
          "created_at"
        ],
        "properties": {
          "allow_request_professional_agreement": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deactivated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "gender": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_searchable": {
            "type": "boolean"
          },
          "last_login_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "lastname": {
            "type": "string"
          },
          "phone_number": {
            "type": [
              "string",
              "null"
            ]
          },
          "pk_driver_id": {
            "type": "string",
            "format": "uuid"
          },
          "verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "UpdateDriverRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "firstname": {
            "type": [
              "string",
              "null"
            ]
          },
          "gender": {
            "type": [
              "string",
              "null"
            ]
          },
          "language": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DriverLanguage"
              }
            ]
          },
          "lastname": {
            "type": [
              "string",
              "null"
            ]
          },
          "password": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone_number": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}